// src/app/configurator-popup/configurator-popup.interface.ts
export interface IConfiguratorPopup {
  id: number;
  code: string;
  price: number;
  priceComplete: number;
  unit: string;
  amount: number,
  availability: string;
  step: number;
  lengthValue: number;
  value: number;
  minValue: number;
  maxValue: number;
  maxValue2: number;
  defaultValue: number;
  selectedValue: number;
  codeA: string;
  unitA: string;
  stepA: number;
  lengthValueA: number;
  valueA: number;
  minValueA: number;
  maxValueA: number;
  defaultValueA: number;
  selectedValueA: number;
  codeB: string;
  unitB: string;
  stepB: number;
  lengthValueB: number;
  valueB: number;
  minValueB: number;
  maxValueB: number;
  defaultValueB: number;
  selectedValueB: number;
};